Update README.md
Updating ENCRYPTION MANAGER

To update ENCRYPTION MANAGER, simply download the latest version of the source code from the GitHub repository and replace the old files with the new ones.
Contributing

Contributions to ENCRYPTION MANAGER are welcome! If you have an idea for a new feature or find a bug, please submit an issue or pull request on the GitHub repository.